backend/requirements.txt
fastapi
uvicorn[standard]
langchain-openai
langchain-core
pillow
requests
python-multipart
chardet
google-auth
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
numpy
scikit-learn
pandas
matplotlib
seaborn
sqlalchemy
mysql-connector-python
alembic